split base class to one file. Delete some unnecessary code in base class
